+++ plugin/plugin/src/main/kotlin/net/twisterrob/gradle/android/AndroidBuildPlugin.kt
@@ -130,8 +130,10 @@ class AndroidBuildPlugin : BasePlugin() {
 			}
 		}
 
-		if (twisterrob.decorateBuildConfig) {
-			decorateBuildConfig()
+		project.beforeAndroidTasksCreated {
+			if (twisterrob.decorateBuildConfig) {
+				decorateBuildConfig()
+			}
 		}
 		project.afterEvaluate {
 			if (project.plugins.hasAndroid()) {