+++ plugin/plugin/src/main/kotlin/net/twisterrob/gradle/android/AndroidLibraryPlugin.kt
@@ -11,6 +11,7 @@ class AndroidLibraryPlugin : BaseExposedPlugin() {
 	override fun apply(target: Project) {
 		super.apply(target)
 
+		project.apply<AndroidLifecyclePlugin>()
 		project.plugins.apply("com.android.library")
 
 		project.apply<VCSPlugin>()